Revert "Merge pull request #46 from kabakmichal/continue-adding-todo"
This reverts commit 3fa44352c52ee2382f69870dd1a75500ebef80bb, reversing
changes made to 58650ede5418e7eb0cf8766683f2cd25ad0db9d1.
