+++ src/Components/EditedTile/EditedTile.module.css
@@ -8,7 +8,6 @@
   flex-direction: column;
   font-family: Arial, Helvetica, sans-serif;
   padding-bottom: 15px;
-  margin-right: 28px;
 }
 
 .top_container {
@@ -35,7 +34,7 @@
   justify-content: center;
   width: 80%;
   border-radius: 0;
-  border-bottom: 1px solid #00d7ff;
+  border-bottom: 1px solid #00D7FF;
 }
 
 .cancel_btn {
@@ -65,18 +64,18 @@
 }
 
 .tile_title_text {
-  font-family: "HelveticaNeueCyr";
-  font-style: normal;
-  font-weight: 700;
-  font-size: 12px;
-  line-height: 12px;
+  font-family: 'HelveticaNeueCyr';
+font-style: normal;
+font-weight: 700;
+font-size: 12px;
+line-height: 12px;
 
-  /* identical to box height */
-  text-align: center;
-  letter-spacing: 0.02em;
-  text-transform: uppercase;
-  margin-top: 20px;
-  color: #b9c3c8;
+/* identical to box height */
+text-align: center;
+letter-spacing: 0.02em;
+text-transform: uppercase;
+margin-top: 20px;
+color: #B9C3C8;
 }
 
 .bottom_container {
@@ -86,6 +85,7 @@
   height: 100%;
   justify-content: center;
   align-items: flex-end;
+
 }
 
 .bottom_row {
@@ -99,17 +99,17 @@
   margin-right: 21px;
 }
 
-.create_btn {
-  font-family: "HelveticaNeueCyr";
-  font-style: normal;
-  font-weight: 400;
-  font-size: 14px;
-  line-height: 14px;
+.create_btn{
+  font-family: 'HelveticaNeueCyr';
+font-style: normal;
+font-weight: 400;
+font-size: 14px;
+line-height: 14px;
 
-  /* identical to box height */
-  text-align: right;
-  letter-spacing: 0.02em;
+/* identical to box height */
+text-align: right;
+letter-spacing: 0.02em;
 
-  color: #00d7ff;
-  padding-right: 21px;
+color: #00D7FF;
+padding-right: 21px;
 }
