+++ src/Components/ListOfTiles/index.jsx
@@ -17,14 +17,20 @@ export const ListOfTiles = () => {
       headers: { Authorization: `Bearer ${token}` },
     });
 
-  const [showComponent, setShowComponent] = useState(false);
-
-  const addComponent = () => {
-    setShowComponent(true);
-  };
-
-  const resetComponent = () => {
-    setShowComponent(false);
+  const addObject = () => {
+    const newObject = {
+      // title: "New quest",
+      // date: "2023-03-07",
+      // type: "Job",
+      title: "",
+      difficulty: "",
+      date: "2023-03-08",
+      time: "23:00:00",
+      status: "undone",
+      category: "",
+      type: "quest",
+    };
+    setObjects([...objects, newObject]);
   };
 
   let todayQuests = [];
@@ -44,9 +50,7 @@ export const ListOfTiles = () => {
   }, []);
 
   if (objects === null) return <div>Loading...</div>;
-  todayQuests = objects.filter(
-    (object) => object.date === today && object.status === "undone"
-  );
+  todayQuests = objects.filter((object) => object.date === today);
   tomorrowQuests = objects.filter(
     (object) => object.date === tomorrow && object.status === "undone"
   );
@@ -54,13 +58,11 @@ export const ListOfTiles = () => {
 
   return (
     <>
-      <AddButton onClick={addComponent}></AddButton>
-
+      <AddButton onButtonClick={addObject}></AddButton>
       <div className={styles.today_section}>
         <p className={styles.today_section_text}>TODAY</p>
         <div className={styles.today_section_cards}>
           <ul className={styles.today_section_list}>
-            {showComponent && <EditedTile onClose={resetComponent} />}
             {todayQuests.map((obj) => (
               <li key={obj._id} className={styles.list_item}>
                 <Tile
